Don't compare columns to character position
This caused TAB to move your cursor to the beginning of the line when
(re)indenting in a file that used tabs for indentation.

(save-excursion (back-to-indentation) (point)) seems expensive.  I
originally compared (current-indentation) to (current-column) instead,
but I found that (current-column) often *lied* when used in conjunction
with smart-tabs-mode, which modifies tab-width inside the python.el
indent functions.  This seemed to be some kind of race, as I would get
the wrong number from (current-column), (sleep-for 1), and then I could
get the right number from (current-column), all
within (python-indent-line).

I fought with that for a couple hours, and eventually landed on just
using character positions instead of column positions.

I also removed the comparison against (line-beginning-position) because,
as far as I can tell, that condition will always be true.