Don't *add* indentation when typing ":"
The code here seems to say that it wants to "[j]ust re-indent
dedenters", so let's make sure we're doing just that.

This was bugging me a lot because I often have already un-indented a
line such as `elif ...` and then I end it with `:`, only to have my
indentation undone.